requirements.txt
Flask==1.1.2
ruamel.yaml==0.16.10
requests==2.23.0
gunicorn==20.0.4
